pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["template_agent"]

[project]
name = "template-agent"
version = "0.1.0"
description = "A template for Model Context Protocol (MCP) server development"
readme = "README.md"
keywords = ["mcp", "template", "server"]
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.6",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asn1crypto==1.5.1",
    "attrs==25.3.0",
    "authlib==1.6.0",
    "backoff==2.2.1",
    "blinker==1.9.0",
    "boto3==1.38.27",
    "botocore==1.38.27",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cryptography==45.0.3",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "fastmcp==2.8.1",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.6.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.24.2",
    "google-auth==2.40.2",
    "googleapis-common-protos==1.70.0",
    "groq==0.26.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.63",
    "langchain-google-genai==2.0.11",
    "langchain-groq==0.2.5",
    "langchain-mcp-adapters==0.1.1",
    "langchain-text-splitters==0.3.8",
    "langfuse==2.60.5",
    "langgraph==0.4.7",
    "langgraph-checkpoint==2.0.26",
    "langgraph-checkpoint-mongodb==0.1.3",
    "langgraph-checkpoint-postgres==2.0.21",
    "langgraph-checkpoint-sqlite==2.0.10",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.43",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mcp==1.9.4",
    "mdurl==0.1.2",
    "motor==3.7.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "narwhals==1.41.0",
    "numpy==2.2.6",
    "oauthlib==3.2.2",
    "openapi-pydantic==0.5.1",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg-pool==3.2.6",
    "psycopg2-binary==2.9.10",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pymongo==4.11.3",
    "pyopenssl==25.1.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "resend==2.8.0",
    "rich==14.0.0",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "s3transfer==0.13.0",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.41",
    "sqlite-vec==0.1.6",
    "sse-starlette==2.3.6",
    "starlette==0.46.2",
    "streamlit==1.45.1",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "toml==0.10.2",
    "tomlkit==0.13.2",
    "tornado==6.5.1",
    "typer==0.16.0",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.32.1",
    "watchdog==6.0.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zstandard==0.23.0",
    "structlog>=24.1.0",
    "snowflake-connector-python>=3.12.0,<4.0",
    "sqlglot>=25.0.0,<26.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "pytest-cov==6.2.1",
    "ruff==0.12.2",
    "mypy==1.16.1",
    "pre-commit==4.2.0",
    "httpx==0.28.1"
]

[project.scripts]
template-agent = "template_agent.src.main:main"

[project.urls]
Homepage = "https://github.com/redhat-data-and-ai/template-agent"
Repository = "https://github.com/redhat-data-and-ai/template-agent"
Issues = "https://github.com/redhat-data-and-ai/template-agent/issues"

[tool.ruff]
target-version = "py312"
line-length = 88
lint.select = ["E", "W", "F", "I", "F401"]
lint.ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
]
markers = ["asyncio"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
